requirements.txt
datasets==2.20.0
python-dotenv==1.0.1
psycopg2==2.9.9
pgcopy==1.6.0